+++ src/Exception/Exception.php
@@ -1,0 +1,15 @@
+<?php
+declare(strict_types=1);
+
+namespace Emailcenter\MaxemailApi\Exception;
+
+/**
+ * Maxemail API Client
+ *
+ * @package    Emailcenter/MaxemailApi
+ * @copyright  2007-2017 Emailcenter UK Ltd. (https://www.emailcenteruk.com)
+ * @license    LGPL-3.0
+ */
+interface Exception
+{
+}