+++ CHANGELOG.md
@@ -6,6 +6,35 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [4.x] - TBC
+### Added
+- Apply GNU LGPLv3 software licence
+- Dedicated package-specific Exception classes. These extend the original SPL
+Exception classes, so no changes should be required in any implementation. This
+now means that all package exceptions implement `Mxm\Exception\Exception`.
+- Use [Guzzle](http://guzzlephp.org/) to handle HTTP connection. Exceptions
+thrown for networking errors (connection timeout, DNS errors, etc.) will be an
+instance of `GuzzleHttp\Exception\RequestException`. This extends
+`\RuntimeException` so any existing catch blocks will continue to function
+without any necessary changes. Errors returned by Maxemail as a response to an
+API call will now throw `Emailcenter\MaxemailApi\Exception\ClientException`
+which extends the same `\RuntimeException` used previously.
+- `Helper::setLogLevel()` can be used to increase the log level from the default
+of *debug*. This might be useful if you want to see the file transfer logs at
+*info* and filter out the raw HTTP client logs at *debug*.
+### Changed
+- Namespaces have been changed from `Mxm\Api` to `Emailcenter\MaxemailApi`.
+- Deprecate API config keys for *user* and *pass*, replaced with *username* and
+*password* respectively
+- Primary class has changed from `Mxm\Api` to `Emailcenter\MaxemailApi\Client`.
+See [README](README.md) for usage example including config changes above.
+### Removed
+- Removed support for PHP 5.x as it is no longer
+[actively supported](https://php.net/supported-versions.php) by the PHP project
+- Remove API config key for *host*. Most implementations using the default
+instance of Maxemail will not need to make any other change. If connecting to a
+non-standard instance of Maxemail, a new key *uri* can be provided. 
+
 ## [3.1.2] - 2017-01-16
 ### Changed
 - Removed API method name from debug log message (still available in log context)